+++ src/main/java/TM/CoordinationOffice/service/impl/ICoordinatorService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface ICoordinatorService {
     public List<Coordinator> getAllCoordinators();
-    public Coordinator getCoordinatorById(int id);
+    public Coordinator getCoordinatorById(long id);
     public int saveOrUpdate(Coordinator coordinator);
-    public void deleteById(int id);
-    public void update(Coordinator coordinator, int id);
+    public void deleteById(long id);
+    public void update(Coordinator coordinator, long id);
 }
